feat: delete specific user task

+++ src/tasks/tasks.service.ts
@@ -16,7 +16,7 @@ export class TasksService {
     }
 
     async getTaskById(id: number, user: User): Promise<Task> {
-        const task = await this.taskRepository.findOne({ where: { id, userId: user.id}});
+        const task = await this.taskRepository.findOne({ where: { id, userId: user.id }});
         
         if (!task) throw new NotFoundException("Task not found!");
         
@@ -36,8 +36,8 @@ export class TasksService {
         return task;
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id)
+    async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id })
         
         if (result.affected === 0) throw new NotFoundException("Task not found!");
     }
